requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Admin
Flask-WTF
PyMySQL
pypdf
pandas
Pillow
WeasyPrint
XlsxWriter
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2   
Flask-Migrate 